+++ poium/__init__.py
@@ -24,4 +24,4 @@ from .javascript import CSSElement
 
 __author__ = "fnngj"
 
-__version__ = "1.1.0"
+__version__ = "1.0.4"